KNOX-2730 - Fix missing Correlation id in audit logs
