+++ gateway-server/src/main/java/org/apache/knox/gateway/GatewayServlet.java
@@ -17,23 +17,17 @@
  */
 package org.apache.knox.gateway;
 
-import org.apache.knox.gateway.audit.api.Action;
-import org.apache.knox.gateway.audit.api.ActionOutcome;
 import org.apache.knox.gateway.audit.api.AuditService;
 import org.apache.knox.gateway.audit.api.AuditServiceFactory;
-import org.apache.knox.gateway.audit.api.Auditor;
-import org.apache.knox.gateway.audit.api.ResourceType;
-import org.apache.knox.gateway.audit.log4j.audit.AuditConstants;
+import org.apache.knox.gateway.audit.api.CorrelationServiceFactory;
 import org.apache.knox.gateway.config.GatewayConfig;
 import org.apache.knox.gateway.descriptor.GatewayDescriptor;
 import org.apache.knox.gateway.descriptor.GatewayDescriptorFactory;
-import org.apache.knox.gateway.filter.AbstractGatewayFilter;
 import org.apache.knox.gateway.i18n.messages.MessagesFactory;
 import org.apache.knox.gateway.i18n.resources.ResourcesFactory;
 import org.apache.knox.gateway.services.ServiceType;
 import org.apache.knox.gateway.services.GatewayServices;
 import org.apache.knox.gateway.services.metrics.MetricsService;
-import org.apache.knox.gateway.util.ServletRequestUtils;
 
 import javax.servlet.Filter;
 import javax.servlet.FilterChain;
@@ -60,8 +54,6 @@ public class GatewayServlet implements Servlet, Filter {
   private static final GatewayMessages LOG = MessagesFactory.get( GatewayMessages.class );
 
   private static final AuditService auditService = AuditServiceFactory.getAuditService();
-  private static final Auditor auditor = AuditServiceFactory.getAuditService().getAuditor(AuditConstants.DEFAULT_AUDITOR_NAME, AuditConstants.KNOX_SERVICE_NAME,
-      AuditConstants.KNOX_COMPONENT_NAME);
 
   private FilterConfigAdapter filterConfig;
   private GatewayFilter filter;
@@ -140,9 +132,10 @@ public class GatewayServlet implements Servlet, Filter {
       } else {
         ((HttpServletResponse)servletResponse).setStatus( HttpServletResponse.SC_SERVICE_UNAVAILABLE );
       }
-      auditLog(servletRequest, servletResponse);
     } finally {
       auditService.detachContext();
+      // Make sure to destroy the correlationContext to prevent threading issues
+      CorrelationServiceFactory.getCorrelationService().detachContext();
     }
   }
 
@@ -168,24 +161,13 @@ public class GatewayServlet implements Servlet, Filter {
       } else {
         ((HttpServletResponse)servletResponse).setStatus( HttpServletResponse.SC_SERVICE_UNAVAILABLE );
       }
-      auditLog(servletRequest, servletResponse);
     } finally {
       auditService.detachContext();
+      // Make sure to destroy the correlationContext to prevent threading issues
+      CorrelationServiceFactory.getCorrelationService().detachContext();
     }
   }
 
-  private void auditLog(ServletRequest servletRequest, ServletResponse servletResponse) {
-    final int status = ((HttpServletResponse) servletResponse).getStatus();
-    final String requestUri, actionOutcome;
-    if (HttpServletResponse.SC_SERVICE_UNAVAILABLE == status) {
-      requestUri = ServletRequestUtils.getContextPathWithQuery(servletRequest);
-      actionOutcome = ActionOutcome.UNAVAILABLE;
-    } else {
-      requestUri = (String) servletRequest.getAttribute(AbstractGatewayFilter.SOURCE_REQUEST_CONTEXT_URL_ATTRIBUTE_NAME);
-      actionOutcome = ActionOutcome.SUCCESS;
-    }
-    auditor.audit(Action.ACCESS, requestUri, ResourceType.URI, actionOutcome, res.responseStatus(status));
-  }
 
   @Override
   public String getServletInfo() {
