+++ gateway-server/src/test/java/org/apache/knox/gateway/AuditLoggingTest.java
@@ -30,6 +30,7 @@ import org.apache.knox.gateway.audit.log4j.audit.Log4jAuditContext;
 import org.apache.knox.gateway.audit.log4j.correlation.Log4jCorrelationContext;
 import org.apache.knox.gateway.config.GatewayConfig;
 import org.apache.knox.gateway.dispatch.DefaultDispatch;
+import org.apache.knox.gateway.filter.AbstractGatewayFilter;
 import org.apache.knox.test.log.CollectAppender;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -110,6 +111,9 @@ public class AuditLoggingTest {
     EasyMock.expect( request.getServletContext() ).andReturn( context ).anyTimes();
     EasyMock.expect( context.getAttribute(
         GatewayConfig.GATEWAY_CONFIG_ATTRIBUTE)).andReturn(gatewayConfig).anyTimes();
+    EasyMock.expect( request.getAttribute(
+        AbstractGatewayFilter.SOURCE_REQUEST_CONTEXT_URL_ATTRIBUTE_NAME))
+        .andReturn( CONTEXT_PATH+PATH ).anyTimes();
     EasyMock.expect(gatewayConfig.getHeaderNameForRemoteAddress()).andReturn(
         "Custom-Forwarded-For").anyTimes();
     EasyMock.replay( request );
@@ -148,12 +152,14 @@ public class AuditLoggingTest {
     executor.awaitTermination(5, TimeUnit.SECONDS);
     assertThat(executor.isTerminated(), is(true));
 
-    assertThat( CollectAppender.queue.size(), is( numberTotalRequests ) );
+    assertThat( CollectAppender.queue.size(), is( numberTotalRequests * 2) );
 
     // Use a set to make sure to dedupe any requestIds to get only unique ones
     Set<String> requestIds = new HashSet<>();
     for (LogEvent accessEvent : CollectAppender.queue) {
-      verifyAuditEvent( accessEvent, CONTEXT_PATH + PATH, ResourceType.URI, Action.ACCESS, ActionOutcome.UNAVAILABLE, null, "Request method: GET" );
+      verifyAuditEvent( accessEvent, CONTEXT_PATH + PATH, ResourceType.URI,
+          Action.ACCESS, accessEvent.getContextData().getValue("outcome"),
+          null, accessEvent.getMessage().getFormattedMessage() );
 
       CorrelationContext cc = Log4jCorrelationContext.of(accessEvent);
       // There are some events that do not have a CorrelationContext associated (ie: deploy)
@@ -210,7 +216,8 @@ public class AuditLoggingTest {
     gateway.doFilter( request, response, chain );
     gateway.destroy();
 
-    assertThat( CollectAppender.queue.size(), is( 1 ) );
+    // Now there are two audit log messages
+    assertThat( CollectAppender.queue.size(), is( 2 ) );
     Iterator<LogEvent> iterator = CollectAppender.queue.iterator();
     LogEvent accessEvent = iterator.next();
     verifyAuditEvent( accessEvent, CONTEXT_PATH + PATH, ResourceType.URI,
